+++ config/urls.py
@@ -12,7 +12,7 @@ from rest_framework_swagger.views import get_swagger_view
 
 from config.views import robots_txt
 from seed.api.base.urls import urlpatterns as api
-from seed.views.main import angular_js_tests, pm_integration, pm_integration_worker, pm_integration_get_templates
+from seed.views.main import angular_js_tests
 
 urlpatterns = [
     # Application
@@ -28,11 +28,6 @@ urlpatterns = [
     # API
     url(r'^api/swagger/', get_swagger_view(title='SEED API'), name='swagger'),
     url(r'^api/', include(api, namespace='api')),
-
-    # Temporary
-    url(r'^pm_integration/', pm_integration, name='pm_integration'),
-    url(r'^pm_integration_worker/', pm_integration_worker, name='pm_integration_worker'),
-    url(r'^pm_integration_get_templates', pm_integration_get_templates, name='pm_integration_get_templates'),
 ]
 
 handler404 = 'seed.views.main.error404'
