+++ seed/templates/seed/_sidebar.html
@@ -39,9 +39,5 @@
                 <div class="icon"><i class="fa fa-sign-out"></i></div>
                 <div ng-show="!is_initial_state" class="item_name">Log Out</div>
             </a>
-            <a id="sidebar-api" ng-class="{item: true, active: is_active('/pm_integration/')}" title="PM Integration Page" href="{% url "pm_integration" %}">
-                <div class="icon"><i class="fa fa-exchange"></i></div>
-                <div ng-show="!is_initial_state" class="item_name">PM Integration Page</div>
-            </a>
         </div>
     </div>
